fix: [Programmers] #181927 - 불필요한 캐스팅 제거

+++ 2026-08/src/day_24/prms_181927/Solution.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 vector<int> solution(vector<int> num_list) {
-    int n = (int)num_list.size();
+    int n = num_list.size();
 
     if (num_list[n - 1] > num_list[n - 2]) {
         num_list.push_back(num_list[n - 1] - num_list[n - 2]);
